refactor: restructure project folders and refactor files
- Moved files into new folders structure for better modularity
- Refactored controllers, use cases, and repository implementations
- Updated imports to reflect new file locations
- Improved code organization to align with clean architecture principles
