+++ compose.yaml
@@ -1,5 +1,3 @@
-version: "3.8"
-
 services:
   postgresql:
     container_name: clean_arch_postgres
@@ -20,6 +18,7 @@ services:
       retries: 5
 
   pgadmin:
+    container_name: clean_arch_pgadmin
     image: dpage/pgadmin4
     environment:
       PGADMIN_DEFAULT_EMAIL: ${PGADMIN_EMAIL}
