+++ package.json
@@ -8,6 +8,7 @@
     "services:up": "docker compose -f compose.yaml up -d",
     "services:stop": "docker compose -f compose.yaml stop",
     "services:down": "docker compose -f compose.yaml down",
+    "destroy": "docker system prune -a",
     "lint:check": "prettier --check .",
     "lint:fix": "prettier --write .",
     "migration:create": "node-pg-migrate --migrations-dir src/infra/migrations create",
