--- src/application/usecases/product-usecases/CreateProductUseCase.ts
@@ -1,12 +1,0 @@
-import { Product } from "domain/entities/Product";
-import { ProductRepository } from "domain/repository/product/ProductRepository";
-
-export class CreateProductUseCase {
-  constructor(private productRepository: ProductRepository) {
-    this.productRepository = productRepository;
-  }
-
-  async execute(product: Product): Promise<Product> {
-    return await this.productRepository.createItem(product);
-  }
-}
