+++ src/infrastructure/http/routes/rootRoute.ts
