+++ src/shared/abstractions/IProductProps.ts
@@ -1,0 +1,6 @@
+export interface IProductProps {
+  name: string;
+  description: string;
+  price: number;
+  active: boolean;
+}
