+++ src/shared/abstractions/IUseCase.ts
@@ -1,0 +1,3 @@
+export default interface IUseCase<I, O> {
+  execute(i: I): Promise<O>;
+}
